% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_solver}
\alias{example_solver}
\title{Create a HiGHS Solver Object}
\usage{
example_solver(op_type = c("LP", "MILP", "QP"))
}
\arguments{
\item{op_type}{Character string specifying the type of optimization model.
Must be one of "LP", "MILP", or "QP".}
}
\value{
An object of class \code{"highs_solver"}.
}
\description{
Creates and solves an example optimization model using the HiGHS solver.
This is a convenience wrapper that combines model creation and solving
in a single function call.
}
\examples{
solver <- example_solver("LP")
solver <- example_solver("MILP")
solver <- example_solver("QP")

}
