% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_get_ncons}
\alias{hi_model_get_ncons}
\title{Get Number of Constraints in a Model}
\usage{
hi_model_get_ncons(model)
}
\arguments{
\item{model}{A `highs_model` object. The model from which to get the number of variables.}
}
\value{
An integer representing the number of constraints in the model.
}
\description{
This function retrieves the number of constraints in a given `highs_model` object.
}
\examples{
model <- hi_new_model()
hi_model_get_ncons(model)

}
