% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_get_nvars}
\alias{hi_model_get_nvars}
\title{Get Number of Variables in a Highs Model}
\usage{
hi_model_get_nvars(model)
}
\arguments{
\item{model}{A `highs_model` object. The model from which to get the number of variables.}
}
\value{
An integer representing the number of variables in the model.
}
\description{
This function retrieves the number of variables in a given Highs model.
}
\examples{
model <- hi_new_model()
hi_model_get_nvars(model)

}
