% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_lower}
\alias{hi_model_set_lower}
\title{Set Lower Bounds for Highs Model}
\usage{
hi_model_set_lower(model, lower)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{lower}{a numeric vector giving the lower bounds.}
}
\value{
\code{NULL}
}
\description{
This function sets the lower bounds for a given Highs model.
}
\examples{
model <- hi_new_model()
lower_bounds <- c(0, 1, 2)
hi_model_set_lower(model, lower_bounds)

}
