% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_ncol}
\alias{hi_model_set_ncol}
\title{Sets the number of columns in the model}
\usage{
hi_model_set_ncol(model, ncol)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{ncol}{an integer giving the number of columns (variables) to set in the model}
}
\value{
\code{NULL}
}
\description{
This function sets the number of columns in the given model.
}
\examples{
model <- hi_new_model()
hi_model_set_ncol(model, 10L) # Sets the model to have 10 columns
}
