% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_objective}
\alias{hi_model_set_objective}
\title{Set Objective for Highs Model}
\usage{
hi_model_set_objective(model, objective)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{objective}{a numeric vector giving the objective values to be set for the model.}
}
\value{
\code{NULL}
}
\description{
This function sets the objective for a given Highs model.
}
\examples{
model <- hi_new_model()
objective <- c(1, 2, 3)
hi_model_set_objective(model, objective)

}
