% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_offset}
\alias{hi_model_set_offset}
\title{Set Offset for Highs Model}
\usage{
hi_model_set_offset(model, offset)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{offset}{a numeric value of length 1. The offset value to be set for the model.}
}
\value{
\code{NULL}
}
\description{
This function sets the offset for a given Highs model.
}
\examples{
model <- hi_new_model()
hi_model_set_offset(model, 10)

}
