% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_rhs}
\alias{hi_model_set_rhs}
\title{Set Right Hand Side for a Highs Model}
\usage{
hi_model_set_rhs(model, rhs)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{rhs}{a numeric vector giving the left hand side values.}
}
\value{
\code{NULL}
}
\description{
This function sets the left hand side for a given Highs model.
}
\examples{
model <- hi_new_model()
model <- hi_model_set_rhs(model, c(0, 1, 2))

}
