% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_sense}
\alias{hi_model_set_sense}
\title{Set the sense of the optimization model}
\usage{
hi_model_set_sense(model, maximum)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{maximum}{a boolean value indicating whether the model should be set to maximization (`TRUE`) or minimization (`FALSE`).}
}
\value{
\code{NULL}
}
\description{
This function sets the sense of the optimization model to either maximization or minimization.
}
\examples{
model <- hi_new_model()
hi_model_set_sense(model, TRUE) # Set the model to maximization
hi_model_set_sense(model, FALSE) # Set the model to minimization

}
