% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_upper}
\alias{hi_model_set_upper}
\title{Set Upper Bounds for a Highs Model}
\usage{
hi_model_set_upper(model, upper)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{upper}{a numeric vector giving the upper bounds.}
}
\value{
\code{NULL}
}
\description{
This function sets the upper bounds for a given Highs model.
}
\examples{
model <- hi_new_model()
upper_bounds <- c(10, 20, 30)
hi_model_set_upper(model, upper_bounds)

}
