% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_model_set_vartype}
\alias{hi_model_set_vartype}
\title{Set Variable Types in a Highs Model}
\usage{
hi_model_set_vartype(model, types)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}.}

\item{types}{an integer vector specifying the types of the variables.}
}
\value{
The function does not return a value. It modifies the `model` object in place.
}
\description{
This function sets the variable types in a given Highs model.
}
\examples{
model <- hi_new_model()
types <- c(1, 2, 1, 0)
hi_model_set_vartype(model, types)

}
