% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{hi_new_model}
\alias{hi_new_model}
\title{Create new Highs Model}
\usage{
hi_new_model()
}
\value{
an object of class \code{"highs_model"}.
}
\description{
Create a new highs model object.
}
\examples{
model <- hi_new_model() 
}
