% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_reset_global_scheduler}
\alias{hi_reset_global_scheduler}
\title{Reset Global Scheduler}
\usage{
hi_reset_global_scheduler(blocking)
}
\arguments{
\item{blocking}{A logical value indicating whether to wait for completion.}
}
\value{
Invisible NULL.
}
\description{
This function resets the global scheduler used by the solver.
}
\examples{
hi_reset_global_scheduler(TRUE)

}
