% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_add_rows}
\alias{hi_solver_add_rows}
\title{Add Constraints to Model}
\usage{
hi_solver_add_rows(solver, lhs, rhs, start, index, value)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{lhs}{A numeric vector of left-hand side bounds.}

\item{rhs}{A numeric vector of right-hand side bounds.}

\item{start}{An integer vector of starting positions in the sparse matrix.}

\item{index}{An integer vector of column indices.}

\item{value}{A numeric vector of coefficient values.}
}
\value{
The solver instance with new constraints added.
}
\description{
This function adds new constraints (rows) to the optimization model.
}
\examples{
solver <- example_solver()
hi_solver_add_rows(solver, c(-Inf), c(10), c(0, 2), c(0, 1), c(1, 2))

}
