% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_add_vars}
\alias{hi_solver_add_vars}
\title{Add Variables to the Solver}
\usage{
hi_solver_add_vars(solver, lower, upper)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{lower}{A numeric vector of lower bounds for the new variables.}

\item{upper}{A numeric vector of upper bounds for the new variables.}
}
\value{
The solver instance with the new variables added.
}
\description{
This function adds new variables to the solver with specified bounds.
}
\examples{
solver <- example_solver()
hi_solver_add_vars(solver, lower = c(0, 0, 0), upper = c(10, 10, 10))

}
