% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_change_constraint_bounds}
\alias{hi_solver_change_constraint_bounds}
\title{Change Constraint Bounds}
\usage{
hi_solver_change_constraint_bounds(solver, idx, lhs, rhs)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{idx}{An integer vector specifying the constraint indices.}

\item{lhs}{The new left-hand side bound.}

\item{rhs}{The new right-hand side bound.}
}
\value{
The solver instance with updated constraint bounds.
}
\description{
This function updates the bounds of an existing constraint in the model.
}
\examples{
solver <- example_solver()
hi_solver_change_constraint_bounds(solver, 1, -Inf, 100)

}
