% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_change_variable_bounds}
\alias{hi_solver_change_variable_bounds}
\title{Change Variable Bounds}
\usage{
hi_solver_change_variable_bounds(solver, idx, lower, upper)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{idx}{An integer specifying the variable index.}

\item{lower}{The new lower bound.}

\item{upper}{The new upper bound.}
}
\value{
The solver instance with updated bounds.
}
\description{
This function updates the bounds of an existing variable in the model.
}
\examples{
solver <- example_solver()
hi_solver_change_variable_bounds(solver, 1, 0, 10)

}
