% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_clear}
\alias{hi_solver_clear}
\title{Clear All Solver Data}
\usage{
hi_solver_clear(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
The cleared solver instance.
}
\description{
This function clears all data from the solver instance, including the model and solution.
}
\examples{
solver <- example_solver()
hi_solver_clear(solver)

}
