% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_int_option}
\alias{hi_solver_get_int_option}
\title{Get Integer Option Value}
\usage{
hi_solver_get_int_option(solver, key)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{key}{A character string specifying the option name.}
}
\value{
An integer value.
}
\description{
This function retrieves the value of an integer solver option.
}
\examples{
solver <- example_solver()
value <- hi_solver_get_int_option(solver, "log_dev_level")

}
