% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_num_col}
\alias{hi_solver_get_num_col}
\title{Get Number of Variables}
\usage{
hi_solver_get_num_col(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
An integer representing the number of variables.
}
\description{
This function returns the number of variables (columns) in the optimization model.
}
\examples{
solver <- example_solver()
n_vars <- hi_solver_get_num_col(solver)

}
