% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{hi_solver_get_option}
\alias{hi_solver_get_option}
\title{Get a HiGHS Solver Option}
\usage{
hi_solver_get_option(
  solver,
  key,
  type = c("auto", "bool", "integer", "double", "string")
)
}
\arguments{
\item{solver}{A HiGHS solver object of class \code{"highs_solver"}.}

\item{key}{A character string specifying the option name to retrieve.}

\item{type}{Type of the option. Can be one of "auto", "bool", "integer", "double", or "string".
When set to "auto" (default), the function will attempt to determine the type from the available options list.
Specify a type directly if the option is valid but not listed in the available options.}
}
\value{
The value of the specified option with the appropriate type.
}
\description{
Retrieves the value of a specific option from a HiGHS solver instance.
}
\examples{
solver <- example_solver()
hi_solver_get_option(solver, "output_flag")
hi_solver_get_option(solver, "solver", type = "string")
}
