% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{hi_solver_get_options}
\alias{hi_solver_get_options}
\title{Get multiple HiGHS Solver Options}
\usage{
hi_solver_get_options(solver, keys = NULL)
}
\arguments{
\item{solver}{A HiGHS solver object of class \code{"highs_solver"}.}

\item{keys}{A character vector of option names to retrieve.}
}
\value{
A named list of option values with the appropriate types.
}
\description{
Retrieves the values of multiple options from a HiGHS solver instance.
}
\examples{
solver <- example_solver()
hi_solver_get_options(solver, c("output_flag", "solver"))
}
