% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_sense}
\alias{hi_solver_get_sense}
\title{Get the optimization sense of the solver instance.}
\usage{
hi_solver_get_sense(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver" representing the solver instance.}
}
\value{
The optimization sense of the solver instance.
}
\description{
This function returns the optimization sense (e.g., minimization or maximization) of the provided solver instance.
}
\examples{
solver <- example_solver()
hi_solver_get_sense(solver)

}
