% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_solution}
\alias{hi_solver_get_solution}
\title{Get Solution}
\usage{
hi_solver_get_solution(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list containing the solution information.
}
\description{
This function retrieves the solution from the solver after optimization.
}
\examples{
solver <- example_solver()
hi_solver_run(solver)
solution <- hi_solver_get_solution(solver)

}
