% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_str_option}
\alias{hi_solver_get_str_option}
\title{Get String Option Value}
\usage{
hi_solver_get_str_option(solver, key)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{key}{A character string specifying the option name.}
}
\value{
A character string.
}
\description{
This function retrieves the value of a string solver option.
}
\examples{
solver <- example_solver()
value <- hi_solver_get_str_option(solver, "solver")

}
