% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_run}
\alias{hi_solver_run}
\title{Run the Solver}
\usage{
hi_solver_run(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
The solver instance after optimization.
}
\description{
This function executes the optimization solver on the current model.
}
\examples{
solver <- example_solver()
hi_solver_run(solver)

}
