% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_constraint_bounds}
\alias{hi_solver_set_constraint_bounds}
\title{Set constraint bounds for a given constraint.}
\usage{
hi_solver_set_constraint_bounds(solver, index, lower, upper)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{index}{The constraint index.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\value{
The solver instance with updated constraint bounds.
}
\description{
This function sets the lower and upper bounds for a specific constraint.
}
\examples{
solver <- example_solver()
hi_solver_set_constraint_bounds(solver, 1, -Inf, 100)

}
