% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_integrality}
\alias{hi_solver_set_integrality}
\title{Set integrality for a variable in the solver.}
\usage{
hi_solver_set_integrality(solver, index, type)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{index}{An integer specifying the variable index.}

\item{type}{An integer representing the integrality type.}
}
\value{
The solver instance with updated integrality settings.
}
\description{
This function defines whether a variable is categorized as integral or continuous.
}
\examples{
solver <- example_solver()
hi_solver_set_integrality(solver, 1, 1)

}
