% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_offset}
\alias{hi_solver_set_offset}
\title{Set the objective offset for the solver.}
\usage{
hi_solver_set_offset(solver, ext_offset)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{ext_offset}{A numeric value representing the offset.}
}
\value{
The solver instance with the updated offset.
}
\description{
This function sets the objective offset in the solver instance.
}
\examples{
solver <- example_solver()
hi_solver_set_offset(solver, 5.0)

}
