% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_sense}
\alias{hi_solver_set_sense}
\title{Set the optimization sense of the solver instance.}
\usage{
hi_solver_set_sense(solver, maximum)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{maximum}{A boolean indicating whether to set maximization (TRUE) or minimization (FALSE).}
}
\value{
The updated solver instance with the new optimization sense.
}
\description{
This function updates the optimization sense for the given solver instance. Use TRUE for maximization and FALSE for minimization.
}
\examples{
solver <- example_solver()
hi_solver_set_sense(solver, TRUE)

}
