% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_variable_bounds}
\alias{hi_solver_set_variable_bounds}
\title{Set variable bounds for a set of variables.}
\usage{
hi_solver_set_variable_bounds(solver, index, lower, upper)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{index}{The variable index.}

\item{lower}{The lower bound.}

\item{upper}{The upper bound.}
}
\value{
The solver instance with updated variable bounds.
}
\description{
This function sets the lower and upper bounds for a set of variables.
}
\examples{
solver <- example_solver()
hi_solver_set_variable_bounds(solver, 2, 0, 10)

}
