% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_write_basis}
\alias{hi_solver_write_basis}
\title{Write Basis to File}
\usage{
hi_solver_write_basis(solver, filename)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{filename}{A character string specifying the output file path.}
}
\value{
Invisible NULL.
}
\description{
This function writes the current basis information to a file.
}
\examples{
solver <- example_solver()
basis_file <- tempfile(fileext = ".txt")
hi_solver_write_basis(solver, basis_file)

}
