% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highs.R
\name{highs_write_model}
\alias{highs_write_model}
\title{Write a Highs Model to a File}
\usage{
highs_write_model(model, file)
}
\arguments{
\item{model}{an object of class \code{highs_model}.}

\item{file}{a character string giving the filename.}
}
\description{
Write an highs model to file.
}
\examples{
model <- example_model()
model_file <- tempfile(fileext = ".mps")
highs_write_model(model, model_file)

}
