% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_objective}
\alias{hi_solver_set_objective}
\title{Set the objective coefficient for a set of variables.}
\usage{
hi_solver_set_objective(solver, index, coeff)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{index}{The variable index.}

\item{coeff}{A numeric value representing the objective coefficient.}
}
\value{
The solver instance with the updated objective.
}
\description{
This function assigns a coefficient to a variable in the objective function.
}
\examples{
solver <- example_solver()
hi_solver_set_objective(solver, 2, 3.5)

}
